Added flexibility on what downloader choose; added exception
By adding the folder downloader the application is now able
to decide, reading the configuration in the conf file,
whether to download the entire file, or just a portion.

It was from the beginning that the resource was save only partially,
just to show the headers,as requested.

Just changing the code for the part that saves the file
is possible to chose which strategy to use.

Added an exception in case the URL has
an host not found or a resource not found.